requirements.txt
py==1.8.0
pytest==4.3.1
pyyaml==5.1
six==1.10.0
testinfra==1.4.2
